modified:   Dictionary.h 	modified:   flashDict.cpp 	new file:   src/periph/seesaw.cpp 	new file:   src/periph/seesaw.h The 'freq' word - placeholder only.  Part of seesaw support for ainsuForth (experimental).
The freq word is meant to control the base (carrier) frequency's specification
for PWM (pulse-width modulation) pins on the Seesaw breakout device, from
remote (via an utterance over an i2c channel).

The factory default seems to be near 490 Hz -- it would be desirable to
change this to (near) 50 Hz, for use with r/c servo motors and such.

That's the working supposition.  No code here, yett.

-wa1tnr

08 January 2018.